💧 drop requirement for class_id in Detections constructor to make it more flexible
