+++ supervision/__init__.py
@@ -1,6 +1,7 @@
-__version__ = "0.3.1"
+__version__ = "0.3.2"
 
-from supervision.detection.core import BoxAnnotator, Detections
+from supervision.detection.core import Detections
+from supervision.detection.annotate import BoxAnnotator
 from supervision.detection.line_counter import LineZone, LineZoneAnnotator
 from supervision.detection.polygon_zone import PolygonZone, PolygonZoneAnnotator
 from supervision.detection.utils import generate_2d_mask
